# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DriveIQ API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /files/search:
    get:
      operationId: searchFiles
      tags: [files]
      summary: Search files across all Drive locations
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: Search query text
        - name: fileType
          in: query
          schema:
            type: string
          description: Filter by MIME type category (document, spreadsheet, presentation, pdf, image, folder, other)
        - name: owner
          in: query
          schema:
            type: string
          description: Filter by owner email
        - name: modifiedAfter
          in: query
          schema:
            type: string
          description: ISO date string for modified after filter
        - name: modifiedBefore
          in: query
          schema:
            type: string
          description: ISO date string for modified before filter
        - name: location
          in: query
          schema:
            type: string
            enum: [myDrive, sharedDrive, sharedWithMe]
          description: Filter by location
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [relevance, modifiedTime, name]
            default: relevance
          description: Sort order
        - name: pageToken
          in: query
          schema:
            type: string
          description: Pagination token
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
          description: Results per page
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileSearchResult"
  /files/{fileId}:
    get:
      operationId: getFileDetails
      tags: [files]
      summary: Get detailed file metadata
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriveFile"
  /files/{fileId}/path:
    get:
      operationId: getFilePath
      tags: [files]
      summary: Get full folder path to a file
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File path breadcrumbs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilePath"
  /files/{fileId}/permissions:
    get:
      operationId: getFilePermissions
      tags: [files]
      summary: Get permissions breakdown for a file
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Permission details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilePermissions"
  /files/{fileId}/preview-url:
    get:
      operationId: getFilePreviewUrl
      tags: [files]
      summary: Get preview URL for a file
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Preview URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilePreviewUrl"
  /files/{fileId}/export-csv:
    get:
      operationId: exportPermissionsCsv
      tags: [files]
      summary: Export permissions audit as CSV
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: CSV content
          content:
            text/csv:
              schema:
                type: string
  /shared-files:
    get:
      operationId: getSharedFiles
      tags: [shared]
      summary: Get Shared with me files with grouping
      parameters:
        - name: groupBy
          in: query
          schema:
            type: string
            enum: [person, fileType, date]
            default: person
          description: How to group results
        - name: staleOnly
          in: query
          schema:
            type: boolean
            default: false
          description: Show only stale files (not modified in 90+ days)
        - name: search
          in: query
          schema:
            type: string
          description: Search within shared files
        - name: pageToken
          in: query
          schema:
            type: string
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Grouped shared files
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SharedFilesResult"
  /team-members:
    get:
      operationId: getTeamMembers
      tags: [team]
      summary: List team members
      responses:
        "200":
          description: Team member list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamMember"
    post:
      operationId: addTeamMember
      tags: [team]
      summary: Add a team member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamMemberInput"
      responses:
        "201":
          description: Team member added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
  /team-members/{id}:
    delete:
      operationId: deleteTeamMember
      tags: [team]
      summary: Remove a team member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Team member removed
...
  /dashboard/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent file activity
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
...
components:
  schemas:
...
    DriveFile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        mimeType:
          type: string
        iconLink:
          type: ["string", "null"]
        thumbnailLink:
          type: ["string", "null"]
        webViewLink:
          type: ["string", "null"]
        size:
          type: ["string", "null"]
        modifiedTime:
          type: string
        createdTime:
          type: ["string", "null"]
        owners:
          type: array
          items:
            $ref: "#/components/schemas/DriveUser"
        lastModifyingUser:
          type: ["object", "null"]
          properties:
            displayName:
              type: string
            emailAddress:
              type: ["string", "null"]
            photoLink:
              type: ["string", "null"]
        parents:
          type: array
          items:
            type: string
        shared:
          type: boolean
        sharingUser:
          type: ["object", "null"]
          properties:
            displayName:
              type: string
            emailAddress:
              type: ["string", "null"]
            photoLink:
              type: ["string", "null"]
        locationBreadcrumb:
          type: ["string", "null"]
        permissionsSummary:
          type: ["string", "null"]
      required:
        - id
        - name
        - mimeType
        - modifiedTime
    DriveUser:
      type: object
      properties:
        displayName:
          type: string
        emailAddress:
          type: ["string", "null"]
        photoLink:
          type: ["string", "null"]
      required:
        - displayName
    FileSearchResult:
      type: object
      properties:
        files:
          type: array
          items:
            $ref: "#/components/schemas/DriveFile"
        nextPageToken:
          type: ["string", "null"]
        totalCount:
          type: ["integer", "null"]
      required:
        - files
    FilePath:
      type: object
      properties:
        fileId:
          type: string
        fileName:
          type: string
        breadcrumbs:
          type: array
          items:
            $ref: "#/components/schemas/PathSegment"
      required:
        - fileId
        - fileName
        - breadcrumbs
    PathSegment:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
      required:
        - id
        - name
    FilePermissions:
      type: object
      properties:
        fileId:
          type: string
        fileName:
          type: string
        summary:
          type: string
        alertLevel:
          type: string
          enum: [green, amber, red]
        alertMessage:
          type: string
        people:
          type: array
          items:
            $ref: "#/components/schemas/PermissionEntry"
        linkSharing:
          $ref: "#/components/schemas/LinkSharingStatus"
      required:
        - fileId
        - fileName
        - summary
        - alertLevel
        - alertMessage
        - people
        - linkSharing
    PermissionEntry:
      type: object
      properties:
        id:
          type: string
        displayName:
          type: string
        emailAddress:
          type: ["string", "null"]
        photoLink:
          type: ["string", "null"]
        role:
          type: string
        accessOrigin:
          type: string
        grantedDate:
          type: ["string", "null"]
      required:
        - id
        - displayName
        - role
        - accessOrigin
    LinkSharingStatus:
      type: object
      properties:
        enabled:
          type: boolean
        role:
          type: ["string", "null"]
        domain:
          type: ["string", "null"]
        allowEditorsToReshare:
          type: boolean
      required:
        - enabled
        - allowEditorsToReshare
    FilePreviewUrl:
      type: object
      properties:
        fileId:
          type: string
        previewType:
          type: string
          enum: [google, pdf, image, unsupported]
        url:
          type: ["string", "null"]
        downloadUrl:
          type: ["string", "null"]
        mimeType:
          type: string
        fileName:
          type: string
        message:
          type: ["string", "null"]
      required:
        - fileId
        - previewType
        - mimeType
        - fileName
    SharedFilesResult:
      type: object
      properties:
        groups:
          type: array
          items:
            $ref: "#/components/schemas/SharedFileGroup"
        nextPageToken:
          type: ["string", "null"]
      required:
        - groups
    SharedFileGroup:
      type: object
      properties:
        key:
          type: string
        label:
          type: string
        photoUrl:
          type: ["string", "null"]
        files:
          type: array
          items:
            $ref: "#/components/schemas/DriveFile"
        count:
          type: integer
      required:
        - key
        - label
        - files
        - count
    TeamMember:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: ["string", "null"]
        addedAt:
          type: string
      required:
        - id
        - email
        - addedAt
    CreateTeamMemberInput:
      type: object
      properties:
        email:
          type: string
        name:
          type: ["string", "null"]
      required:
        - email
...
    ActivityItem:
      type: object
      properties:
        fileId:
          type: string
        fileName:
          type: string
        action:
          type: string
        actorName:
          type: string
        actorEmail:
          type: ["string", "null"]
        actorPhoto:
          type: ["string", "null"]
        timestamp:
          type: string
        mimeType:
          type: string
      required:
        - fileId
        - fileName
        - action
        - actorName
        - timestamp
        - mimeType
...