# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DriveIQ API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityItem {
  fileId: string;
  fileName: string;
  action: string;
  actorName: string;
  /** @nullable */
  actorEmail?: string | null;
  /** @nullable */
  actorPhoto?: string | null;
  timestamp: string;
  mimeType: string;
}