# lib/api-zod/src/generated/types/createTeamMemberInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DriveIQ API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTeamMemberInput {
  email: string;
  /** @nullable */
  name?: string | null;
}