# lib/api-zod/src/generated/types/driveFile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DriveIQ API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DriveFileLastModifyingUser } from "./driveFileLastModifyingUser";
import type { DriveFileSharingUser } from "./driveFileSharingUser";
import type { DriveUser } from "./driveUser";

export interface DriveFile {
  id: string;
  name: string;
  mimeType: string;
  /** @nullable */
  iconLink?: string | null;
  /** @nullable */
  thumbnailLink?: string | null;
  /** @nullable */
  webViewLink?: string | null;
  /** @nullable */
  size?: string | null;
  modifiedTime: string;
  /** @nullable */
  createdTime?: string | null;
  owners?: DriveUser[];
  /** @nullable */
  lastModifyingUser?: DriveFileLastModifyingUser;
  parents?: string[];
  shared?: boolean;
  /** @nullable */
  sharingUser?: DriveFileSharingUser;
  /** @nullable */
  locationBreadcrumb?: string | null;
  /** @nullable */
  permissionsSummary?: string | null;
}