# lib/api-zod/src/generated/types/filePath.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DriveIQ API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PathSegment } from "./pathSegment";

export interface FilePath {
  fileId: string;
  fileName: string;
  breadcrumbs: PathSegment[];
}