# lib/api-zod/src/generated/types/sharedFileGroup.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DriveIQ API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DriveFile } from "./driveFile";

export interface SharedFileGroup {
  key: string;
  label: string;
  /** @nullable */
  photoUrl?: string | null;
  files: DriveFile[];
  count: number;
}